Add from_chars scenario for iformat in read_doubles.cpp. Why is strtod() the fastest?
